CRUD created

+++ app/Http/Controllers/PostsController.php
@@ -1,12 +1,22 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => [ 'index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -28,6 +38,7 @@ class PostsController extends Controller
     public function create()
     {
         return view('posts.create');
+        
     }
 
     /**
@@ -40,12 +51,35 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle the file upload
+        if($request->hasFile('cover_image')) {
+
+            // get filename with extentions
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store 
+            $fileNameToStore = $filename . '_' .time(). '.' .$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+
+            $fileNameToStore = 'noimage_1588362405.jpeg';
+        }
+
+        
 
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created! Thank you!');
@@ -72,7 +106,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unautorized page');
+        }
         return view('posts.edit')->with('post', $post);
+        
     }
 
     /**
@@ -86,12 +124,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
-        $post =Post::find($id);
+
+        // Handle the file upload
+        if($request->hasFile('cover_image')) {
+
+            // get filename with extentions
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store 
+            $fileNameToStore = $filename . '_' .time(). '.' .$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } 
+
+        $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post updated! Thank you!');
@@ -105,7 +163,16 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
-        $post =Post::find($id);
+         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unautorized page');
+        }
+
+        if($post->cover_image != 'noimage_1588362405.jpeg') {
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post removed!');
     }
